Move location id and url to TraceLocation and create them lazily
