+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/eventregistration/checkins/qrcode/TraceLocation.kt
@@ -1,11 +1,15 @@
 package de.rki.coronawarnapp.eventregistration.checkins.qrcode
 
 import android.os.Parcelable
+import com.google.common.io.BaseEncoding
 import de.rki.coronawarnapp.eventregistration.storage.entity.TraceLocationEntity
 import de.rki.coronawarnapp.server.protocols.internal.pt.TraceLocationOuterClass
+import de.rki.coronawarnapp.ui.eventregistration.organizer.details.QrCodeGenerator
+import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
 import okio.ByteString
 import okio.ByteString.Companion.decodeBase64
+import okio.ByteString.Companion.toByteString
 import org.joda.time.Instant
 
 @Parcelize
@@ -22,6 +26,29 @@ data class TraceLocation(
     val version: Int = VERSION,
 ) : Parcelable {
 
+    /**
+     * Return a url for [TraceLocation] to be used as an input for [QrCodeGenerator]
+     * URL format https://e.coronawarn.app?v=1#QR_CODE_PAYLOAD_BASE64URL
+     */
+    @IgnoredOnParcel
+    val locationUrl: String by lazy {
+        val payloadBytes = qrCodePayload().toByteArray()
+        val base64Url = BaseEncoding.base64Url().omitPadding().encode(payloadBytes)
+        AUTHORITY.plus(base64Url)
+    }
+
+    /**
+     *  Returns a byte sequence that serves as an identifier for the trace location.
+     *  The ID is the byte representation of SHA-256 hash.
+     */
+    @IgnoredOnParcel
+    val locationId: ByteString by lazy {
+        val cwaDomain = CWA_GUID.toByteArray()
+        val payloadBytes = qrCodePayload().toByteArray()
+        val totalByteSequence = cwaDomain + payloadBytes
+        totalByteSequence.toByteString().sha256()
+    }
+
     fun isBeforeStartTime(now: Instant): Boolean = startDate?.isAfter(now) ?: false
 
     fun isAfterEndTime(now: Instant): Boolean = endDate?.isBefore(now) ?: false
@@ -31,6 +58,9 @@ data class TraceLocation(
          * Trace location version. This is a static data and not calculated from [TraceLocation]
          */
         const val VERSION = 1
+
+        private const val AUTHORITY = "https://e.coronawarn.app?v=$VERSION#"
+        private const val CWA_GUID = "CWA-GUID"
     }
 }
